Made the garbage can delete the actual proper warehouse
