+++ src/components/WarehouseList/WarehouseList.js
@@ -1,12 +1,9 @@
-import chevron from "../../assets/Icons/chevron_right-24px.svg";
-import edit from "../../assets/Icons/edit-24px.svg";
-import del from "../../assets/Icons/delete_outline-24px.svg";
 import sortBtn from "../../assets/Icons/sort-24px.svg";
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import axios from "axios";
 
-import DeleteWarehouse from "../DeleteWarehouse/DeleteWarehouse";
+import WarehouseRow from "../WarehouseRow/WarehouseRow";
 
 import "./WarehouseList.scss";
 
@@ -31,15 +28,6 @@ const WarehouseList = () => {
 	const handleSort = () => {
 		warehouses.sort();
 	};
-
-	const dialogRef = useRef(null);
-
-	const toggleModal = () => {
-		if (!dialogRef.current) return;
-		dialogRef.current.hasAttribute("open")
-		? dialogRef.current.close()
-		: dialogRef.current.showModal();
-	}
 
 	if (!loading)
 		return (
@@ -79,44 +67,9 @@ const WarehouseList = () => {
 				{warehouses.map((warehouse) => {
 					return (
 						<>
-							<section className="warehouse-container">
-								<div className="warehouse-container__item">
-									<h4 className="mobile">WAREHOUSE</h4>
-									<span>
-										<Link to={`/warehouses/${warehouse.id}`}>
-											<h3>{warehouse.warehouse_name}</h3>
-										</Link>
-										<img src={chevron} alt="" />
-									</span>
-								</div>
-								<div className="warehouse-container__item">
-									<h4 className="mobile">CONTACT NAME</h4>
-									<p className="p2">{warehouse.contact_name}</p>
-								</div>
-								<div className="warehouse-container__item">
-									<h4 className="mobile">ADDRESS</h4>
-									<p className="p2">
-										{warehouse.address}, {warehouse.city}, {warehouse.country}
-									</p>
-								</div>
-								<div className="warehouse-container__item">
-									<h4 className="mobile">CONTACT INFORMATION</h4>
-									<p className="p2">{warehouse.contact_phone}</p>
-									<p className="p2">{warehouse.contact_email}</p>
-								</div>
-								<div className="warehouse-container__item--col5">
-									<img src={del} alt="Delete" onClick={toggleModal} />
-									<Link to={`/warehouses/${warehouse.id}/edit`}>
-										<img src={edit} alt="Edit" />
-									</Link>
-								</div>
-								<DeleteWarehouse 
-									warehouse_name={warehouse.warehouse_name} 
-									id={warehouse.id} 
-									toggleModal={toggleModal} 
-									ref={dialogRef} 
-								/>
-							</section>
+							<WarehouseRow 
+								warehouse={warehouse}
+							/>
 							<span className="divider"></span>
 						</>
 					);
